Refactor: ts마이그레이션(itemDto)
